+++ packages/analytics/analytics-geo-map/vite.config.ts
@@ -9,6 +9,9 @@ const sanitizedPackageName = sanitizePackageName(packageName)
 
 // Merge the shared Vite config with the local one defined below
 const config = mergeConfig(sharedViteConfig, defineConfig({
+  optimizeDeps: {
+    exclude: ['maplibre-gl'],
+  },
   build: {
     lib: {
       // The kebab-case name of the exposed global variable. MUST be in the format `kong-ui-public-{package-name}`