+++ packages/entities/entities-plugins/src/components/free-form/plugins/datakit/flow-editor/node-forms/PropertiesField.vue
@@ -58,7 +58,7 @@
 <script setup lang="ts">
 import type { SelectItem } from '@kong/kongponents'
 import useI18n from '../../../../../../composables/useI18n'
-import { useFormShared } from '../../../../shared/composables'
+import { useFormShared } from '../../../../core/composables'
 import { computed, inject, ref, watch } from 'vue'
 import { FEATURE_FLAGS } from '../../../../../../constants'
 import {