+++ packages/entities/entities-plugins/src/components/free-form/plugins/rate-limiting-advanced/RateLimitingAdvancedForm.vue
@@ -22,11 +22,11 @@ import { AUTOFILL_SLOT, AUTOFILL_SLOT_NAME } from '@kong-ui-public/forms'
 import { provide } from 'vue'
 import ConfigForm from './ConfigForm.vue'
 import CustomKeyField from './CustomKeyField.vue'
-import FieldRenderer from '../../shared/FieldRenderer.vue'
-import DynamicLayout from '../../shared/layout/DynamicLayout.vue'
+import FieldRenderer from '../../core/components/FieldRenderer.vue'
+import DynamicLayout from '../../layout/DynamicLayout.vue'
 import { useExpressionMode } from '../_shared/use-expression-mode'
 
-import type { PluginFormLayoutProps as Props } from '../../shared/layout/provider'
+import type { PluginFormLayoutProps as Props } from '../../layout/provider'
 import type { FreeFormPluginData } from '../../../../types/plugins/free-form'
 
 const props = defineProps<Props>()