+++ packages/entities/entities-plugins/src/components/free-form/plugins/request-callout/CalloutsForm.vue
@@ -18,8 +18,8 @@
 
 <script setup lang="ts">
 import { getCalloutId } from './utils'
-import { useFormShared } from '../../shared/composables'
-import ArrayField from '../../shared/ArrayField.vue'
+import { useFormShared } from '../../core/composables'
+import ArrayField from '../../core/components/ArrayField.vue'
 import CalloutForm from './CalloutForm.vue'
 
 import { type RequestCalloutPlugin, type Callout, CalloutId } from './types'