+++ packages/core/page-layout/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.9](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fpage-layout%401.1.8...%40kong-ui-public%2Fpage-layout%401.1.9) (2026-09-09)
+
+### Bug Fixes
+
+* **deps:** update all non-major dependencies with stable versions ([#3790](https://github.com/Kong/public-ui-components/issues/3790)) ([aecf1ae](https://github.com/Kong/public-ui-components/commit/aecf1ae1184b73032c10eb89ddc6b9505ba20e63))
+
+
 ## [1.1.8](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fpage-layout%401.1.7...%40kong-ui-public%2Fpage-layout%401.1.8) (2026-09-05)
 
 **Note:** Version bump only for package @kong-ui-public/page-layout